Consolidate cache_serve/_linux and move split into LotMan
